fix: deduplicate @auth/core types and move augmentation to .d.ts
- Move Session/JWT module augmentation from src/lib/auth.ts into a
  dedicated src/types/auth.d.ts so it is picked up across the project.
- Add tsconfig paths mapping @auth/core to ./node_modules/@auth/core
  to deduplicate the type identity (the SDK is linked via file: which
  otherwise loads two copies of @auth/core).
