+++ src/types/auth.d.ts
@@ -1,0 +1,22 @@
+// noinspection JSUnusedGlobalSymbols
+
+import '@auth/core/types';
+import '@auth/core/jwt';
+
+declare module '@auth/core/types' {
+  interface Session {
+    idToken?: string;
+    accessToken?: string;
+    error?: string;
+  }
+}
+
+declare module '@auth/core/jwt' {
+  interface JWT {
+    idToken?: string;
+    accessToken?: string;
+    refreshToken?: string;
+    expiresAt?: number;
+    error?: string;
+  }
+}
